+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * [#141](https://github.com/adamcaudill/yawast/issues/141) - Flag --internalssl as Deprecated
 * [#147](https://github.com/adamcaudill/yawast/issues/147) - User Enumeration via Password Reset Form
 * [#148](https://github.com/adamcaudill/yawast/issues/148) - Added `--vuln_scan` option to enable new vulnerability scanner
+* [#151](https://github.com/adamcaudill/yawast/issues/151) - User Enumeration via Password Reset Form Timing Differences
 * [#130](https://github.com/adamcaudill/yawast/issues/130) - Bug: HSTS Error leads to printing HTML
 * [#132](https://github.com/adamcaudill/yawast/issues/132) - Bug: Typo in SSL Output
 * [#142](https://github.com/adamcaudill/yawast/issues/142) - Bug: Error In Collecting DNS Information